# doc/openapi.yaml
swagger: '2.0'
info:
  contact: {}
  title: Breezy API
  description: API agrégée de la plateforme Breezy (réseau social microservices).
  version: '1.0'
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - auth
      summary: Rafraîchir les tokens
      parameters:
      - description: Refresh token
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      responses:
        '200':
          description: 'Nouvelle paire tokens — data: {token, refresh_token, user}'
          schema:
            $ref: '#/definitions/models.AuthUser'
        '401':
          description: Token absent/invalide/expiré
          schema:
            type: object
            additionalProperties:
              type: string
        '403':
          description: Compte désactivé
          schema:
            type: object
            additionalProperties:
              type: string
        '500':
          description: Erreur interne
          schema:
            type: object
            additionalProperties:
              type: string
  /auth/register:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - auth
      summary: Créer un compte
      parameters:
      - description: Email + mot de passe (min 8 chars)
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      responses:
        '201':
          description: 'Compte créé — data: {token, refresh_token, user}'
          schema:
            $ref: '#/definitions/models.AuthUser'
        '400':
          description: Payload invalide
          schema:
            type: object
            additionalProperties:
              type: string
        '409':
          description: Email déjà utilisé
          schema:
            type: object
            additionalProperties:
              type: string
        '500':
          description: Erreur interne
          schema:
            type: object
            additionalProperties:
              type: string
  /auth/validate:
    get:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - auth
      summary: Valider un access token (usage gateway)
      responses:
        '200':
          description: 'Claims — data: {user_id, email, role}'
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Token invalide
          schema:
            type: object
            additionalProperties:
              type: string
        '500':
          description: Erreur interne
          schema:
            type: object
            additionalProperties:
              type: string
...
  /messages/keys:
    put:
      security:
      - BearerAuth: []
      consumes:
      - application/json
      produces:
      - application/json
      tags:
      - messages
      summary: Publier / mettre à jour sa clé publique E2EE
      parameters:
      - description: Clé publique (base64)
        name: body
        in: body
        required: true
        schema:
          $ref: '#/definitions/models.PublishKeyRequest'
      responses:
        '200':
          description: user_id + public_key
          schema:
            type: object
            additionalProperties:
              type: string
        '400':
          description: Bad Request
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
  /messages/keys/{userId}:
    get:
      security:
      - BearerAuth: []
      produces:
      - application/json
      tags:
      - messages
      summary: Récupérer la clé publique d'un utilisateur
      parameters:
      - type: string
        description: User ID
        name: userId
        in: path
        required: true
      responses:
        '200':
          description: public_key
          schema:
            type: object
            additionalProperties:
              type: string
        '401':
          description: Unauthorized
          schema:
            type: object
            additionalProperties:
              type: string
        '404':
          description: Not Found
          schema:
            type: object
            additionalProperties:
              type: string
...
  /messages/ws:
    get:
      tags:
      - messages
      summary: Connexion WebSocket temps réel (messages)
      parameters:
      - type: string
        description: JWT access token
        name: access_token
        in: query
        required: true
      responses:
        '101':
          description: Upgrade WebSocket
          schema:
            type: string
...
definitions:
  models.AuthUser:
    type: object
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
...
  models.RefreshRequest:
    type: object
    properties:
      refresh_token:
        type: string
  models.RegisterRequest:
    type: object
    required:
    - email
    - password
    properties:
      email:
        type: string
      password:
        type: string
        minLength: 8
...
  models.PublishKeyRequest:
    type: object
    required:
    - public_key
    properties:
      public_key:
        type: string
...
host: localhost:8080
basePath: /